Fixed short format and forced UpdateIndicators on format switching

+++ Source/EngineDecay/EngineDecay/Lib.cs
@@ -12,7 +12,7 @@ namespace EngineDecay
         {
             if (formatCode == 0)
             {
-                return seconds + "s";
+                return (int)seconds + "s";
             }
 
             else if (formatCode == 1)
